% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_benefit_report.R
\name{export_benefit_report}
\alias{export_benefit_report}
\title{export benefit report}
\usage{
export_benefit_report(
  path = ".",
  records_path = "cpd-records",
  report_year = year(today())
)
}
\arguments{
\item{path}{Location for saving the report}

\item{records_path}{Path to the cpd-records}

\item{report_year}{The year for the benefit report}
}
\description{
export benefit report
}
