% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cpd.R
\name{import_cpd}
\alias{import_cpd}
\title{import_cpd}
\usage{
import_cpd(path = "cpd-records")
}
\arguments{
\item{path}{Path to CPD records}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
import_cpd
}
