% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cpd.R
\name{title_to_file}
\alias{title_to_file}
\title{title to file}
\usage{
title_to_file(title)
}
\arguments{
\item{string}{Activity title to be tidied}
}
\value{
Tidy version of activity title
}
\description{
title to file
}
