% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cpd.R
\name{new_cpd}
\alias{new_cpd}
\title{import_cpd}
\usage{
new_cpd(
  title = "",
  type = "other",
  date = today(),
  learning_hours = 0,
  path = "cpd-records",
  overwrite = FALSE
)
}
\arguments{
\item{title}{Title for new CPD record}

\item{type}{Type of new CPD record, can be one of "work", "professional",
"formal", "self" or "other"}

\item{date}{The start date of the activity in yyyy-mm-dd format}

\item{learning_hours}{The number of learning hours acruded during this
activity}

\item{path}{Path to CPD records}

\item{overwrite}{Should the YAML be overwritten if it exists - default FALSE}
}
\description{
import_cpd
}
